+++ sdk/resourcemanager/iothub/armiothub/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Release History
 
+## 1.2.0-beta.1 (2023-04-28)
+### Features Added
+
+- New value `RoutingSourceDigitalTwinChangeEvents`, `RoutingSourceMqttBrokerMessages` added to enum type `RoutingSource`
+- New enum type `IPVersion` with values `IPVersionIPv4`, `IPVersionIPv4IPv6`, `IPVersionIPv6`
+- New struct `EncryptionPropertiesDescription`
+- New struct `KeyVaultKeyProperties`
+- New struct `PropertiesDeviceStreams`
+- New struct `RootCertificateProperties`
+- New struct `RoutingCosmosDBSQLAPIProperties`
+- New field `DeviceStreams` in struct `Properties`
+- New field `Encryption` in struct `Properties`
+- New field `IPVersion` in struct `Properties`
+- New field `RootCertificate` in struct `Properties`
+- New field `CosmosDBSQLCollections` in struct `RoutingEndpoints`
+
+
 ## 1.1.0 (2023-03-31)
 ### Features Added